+++ src/main/java/insane96mcp/iguanatweaksreborn/module/movement/BetterClimbable.java
@@ -2,7 +2,6 @@ package insane96mcp.iguanatweaksreborn.module.movement;
 
 import insane96mcp.iguanatweaksreborn.module.Modules;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
@@ -18,27 +17,23 @@ import net.minecraftforge.event.TickEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModList;
 
-@Label(name = "Better Climbable", description = "Players's slides down climbable blocks faster and stands still when opening an interface. This is disabled if quark is enabled")
-@LoadFeature(module = Modules.Ids.MOVEMENT)
+@LoadFeature(module = Modules.Ids.MOVEMENT, description = "Players's slides down climbable blocks faster and stands still when opening an interface")
 public class BetterClimbable extends Feature {
 
-	@Config(min = 0, max = 5d)
-	@Label(name = "Climb Speed", description = "Speed at which players move up climbable blocks. Vanilla is 0.2")
+	@Config(min = 0, max = 5d, description = "Speed at which players move up climbable blocks. Vanilla is 0.2")
 	public static Double climbSpeed = 0.3d;
-	@Config(min = 0, max = 5d)
-	@Label(name = "Descend Speed", description = "How much faster players move down climbable blocks")
+	@Config(min = 0, max = 5d, description = "How much faster players move down climbable blocks. This is disabled if quark is present")
 	public static Double descendSpeed = 0.2d;
+	@Config(min = 0, max = 5d, description = "If true, and not looking down, player will crouch when opening an interface. This is disabled if quark is present")
+	public static Boolean crouchWhenOpeningInterface = true;
 
-	@Config
-	@Label(name = "Not on climbable when on ground", description = "Entities will not count as on climbable when on the ground, preventing slowdown when passing through climbable blocks.")
+	@Config(description = "Entities will not count as on climbable when on the ground, preventing slowdown when passing through climbable blocks.")
 	public static Boolean notOnClimbableWhenOnGround = true;
 
-	@Config
-	@Label(name = "Only climb with jump", description = "If enabled you'll only be able to climb when pressing jump and not when against a wall and moving.")
+	@Config(description = "If enabled you'll only be able to climb when pressing jump and not when against a wall and moving.")
 	public static Boolean onlyClimbWithJump = true;
 
-	@Config
-	@Label(description = "If enabled, climbables will no longer cancel fall damage. Damage is reduced by 40%.")
+	@Config(description = "If enabled, climbables will no longer cancel fall damage. Damage is reduced by 40%.")
 	public static Boolean fallDamageOnClimbables = true;
 
 	public BetterClimbable(Module module, boolean enabledByDefault, boolean canBeDisabled) {
@@ -70,7 +65,8 @@ public class BetterClimbable extends Feature {
 	@OnlyIn(Dist.CLIENT)
 	public void onInput(MovementInputUpdateEvent event) {
 		if(!this.isEnabled()
-				|| ModList.get().isLoaded("quark"))
+				|| ModList.get().isLoaded("quark")
+				|| !crouchWhenOpeningInterface)
 			return;
 
 		Player player = event.getEntity();